requirements.txt
Django==2.2.4
psycopg2==2.8.3
gunicorn==19.9.0
djangorestframework==3.11.0
dj-static==0.0.6
django-environ==0.4.5